recycling_detection_fastapi/fastapi_app/requirements.txt
torch
torchvision
numpy<2.0
Pillow
supervision
rfdetr
fastapi
uvicorn[standard]
python-multipart
gunicorn